requirements.txt
SQLAlchemy==2.0.0
fastapi==0.89.1
pydantic~=1.10.4
uvicorn~=0.20.0
redis~=4.4.2
hiredis~=2.2.1
pyjwt~=2.6.0
asyncpg~=0.27.0
aiosqlite~=0.18.0
aiosmtplib~=2.0.1
python-multipart~=0.0.5
python-consul~=1.1.0
aiohttp~=3.8.3
aiodns~=3.0.0
python-dotenv~=0.21.1
aiobotocore~=2.4.2
httpx~=0.23.3
pytest~=7.2.1
